Agrega macro-grupo Indicadores de Gestion (50/30/20) y seccion Compromisos; incluye estado acumulado completo de AsignacionService y ResultadoService (resuelve tambien la duplicacion de Trabajo en Equipo)

+++ app/src/Alianzagrafica.Evaluacion180.Web/Models/Entidades/RespuestaEvaluacion.cs
@@ -23,7 +23,11 @@ public class RespuestaDetalle
 {
     public int IdRespuesta { get; set; }
     public int IdCompetencia { get; set; }
-    public byte Calificacion { get; set; }
+
+    /// <summary>Calificación del evaluador, en puntos porcentuales, 0-100 (Entregable 12 — antes
+    /// era una escala de 1 a 5; ver <see cref="Entidades.EscalaCalificacion"/> para la
+    /// clasificación cualitativa Deficiente/Aceptable/Bueno/Sobresaliente equivalente).</summary>
+    public decimal Calificacion { get; set; }
     public string? Comentario { get; set; }
 
     public RespuestaEvaluacion Respuesta { get; set; } = null!;
@@ -32,14 +36,16 @@ public class RespuestaDetalle
 
 /// <summary>Respuesta de un indicador de gestión dentro de una evaluación (Entregable 11),
 /// análoga a <see cref="RespuestaDetalle"/> pero con Meta/Resultado del mes en vez de una
-/// calificación de 1 a 5 — ver <see cref="Entidades.IndicadorGestion"/>.</summary>
+/// calificación directa — ver <see cref="Entidades.IndicadorGestion"/>.</summary>
 public class RespuestaIndicadorDetalle
 {
     public int IdRespuesta { get; set; }
     public int IdIndicador { get; set; }
 
-    /// <summary>Meta del indicador para el periodo, en puntos porcentuales (90 = 90%) —
-    /// informativa (columna "META" del Excel origen); no participa en el cálculo de la nota.</summary>
+    /// <summary>Copia (snapshot) de <see cref="IndicadorGestion.Meta"/> al momento de guardar la
+    /// respuesta, en puntos porcentuales (90 = 90%) — informativa; no participa en el cálculo de
+    /// la nota. Desde el Entregable 12 la Meta es un valor fijo del catálogo que el evaluador ya
+    /// no escribe: el controlador la toma de <see cref="IndicadorGestion"/> al guardar.</summary>
     public decimal? Meta { get; set; }
 
     /// <summary>Resultado real del periodo, en puntos porcentuales (75 = 75%) — es el valor que
